refactor(SinglePokemon) created seperate component for pokemon card
